# Improve JsonSerializer.ReadAsync(Stream) throughput