@Tornhoof 2019-03-06
If you're optimizing for MemoryStream cases, you could also look at ``TryGetBuffer``. It needs to be allowed specifically during creation of the memorystream, but then something like 
```csharp
if (stream is MemoryStream ms && ms.TryGetBuffer(out var buffer))
{
// use buffer directly
}
```
works.