@stephentoub 2019-03-07
> If you're optimizing for MemoryStream cases, you could also look at TryGetBuffer. It needs to be allowed specifically during creation of the memorystream, but then something like

Yup, that might be reasonable.  There's a cost associated with special-casing like that, and a caller could just do it themselves, but if we think MemoryStream would commonly be used here, it'd be reasonable (I don't know if it would be).  I was using it here just as a data point to compare against.